apps/image-processor/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
rembg[cpu]>=2.0.59
pillow>=10.4.0
